requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-chroma>=0.1.0
langchain-experimental>=0.0.50
langgraph>=0.0.28
openai>=1.12.0
chromadb==0.4.15
tiktoken>=0.5.2
unstructured==0.12.4
PyMuPDF==1.23.22
pydantic==2.6.1
lxml==5.1.0
opencv-python==4.9.0.80
pillow==10.2.0
SpeechRecognition==3.10.1
pyttsx3==2.90
gtts==2.5.1
pygame==2.5.2
pyaudio==0.2.14
bleak==0.21.1
keyboard==0.13.5
pandas==2.2.0
numpy==1.26.4
matplotlib==3.8.3
markdown2==2.4.12
beautifulsoup4==4.12.3
requests==2.31.0
nest-asyncio==1.6.0
python-dotenv==1.0.1
flask==3.0.0
flask-socketio==5.3.6
pytest==8.0.0
jupyter==1.0.0
ipykernel==6.29.0
openpyxl
